Changes to Retina rendering
It doesn't really work yet, however it works better now...

+++ src/main/java/cz/krejcar25/sudoku/WinOverlay.java
@@ -7,7 +7,7 @@ public class WinOverlay extends BaseOverlay {
     private Clock clock;
 
     WinOverlay(GameView baseView) {
-        super(baseView, PApplet.constrain((baseView.width - 540) / 2, 0, baseView.width - 540), (baseView.game.rows() * baseView.game.sy / 2) - 100, PApplet.constrain(baseView.width, 0, 540), 200, OverlayType.OK, (sender) -> {
+        super(baseView, PApplet.constrain((baseView.width - 540) / 2, 0, baseView.width - 540), (int)(baseView.game.rows() * baseView.game.sy / 2) - 100, PApplet.constrain(baseView.width, 0, 540), 200, OverlayType.OK, (sender) -> {
             System.out.println("WinOverlay OK button has received click event, popping");
             baseView.viewStack.pop(2);
         });
